[refactor] SwaggerController 형식 맞추기 #67
